9.14

+++ lib/kernel/bitmap.c
@@ -11,7 +11,7 @@ void bitmap_init(struct bitmap* btmp){
 bool bitmap_scan_test(struct bitmap* btmp,uint32_t bit_idx){
     uint32_t byte_idx = bit_idx / 8;
     uint32_t bit_odd = bit_idx % 8;
-    return (btmp->bits[byte_idx] & (uint8_t)(BITMAP_MASK << bit_odd));
+    return (btmp->bits[byte_idx] & (BITMAP_MASK << bit_odd));
 }
 
 int bitmap_scan(struct bitmap* btmp,uint32_t cnt){
